+++ angular/src/app/modules/account/account-review/account-review.component.spec.ts
@@ -1,0 +1,32 @@
+import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { AccountReviewComponent } from './account-review.component';
+
+describe('AccountReviewComponent', () => {
+  const review = {
+    id: '',
+    accountId: '',
+    hotelId: '',
+    comment: '',
+    dateCreated: new Date(),
+    rating: 0,
+  };
+  let component: AccountReviewComponent;
+  let fixture: ComponentFixture<AccountReviewComponent>;
+
+  beforeEach(async(() => {
+    TestBed.configureTestingModule({
+      declarations: [AccountReviewComponent],
+    }).compileComponents();
+  }));
+
+  beforeEach(() => {
+    fixture = TestBed.createComponent(AccountReviewComponent);
+    component = fixture.componentInstance;
+    component.review = review;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});