+++ angular/src/app/modules/account/account-review/account-review.component.ts
@@ -1,0 +1,14 @@
+import { Component, Input, OnInit } from '@angular/core';
+import { Review } from '../../../data/review.model';
+
+@Component({
+  selector: 'uic-account-review',
+  templateUrl: './account-review.component.html',
+})
+export class AccountReviewComponent implements OnInit {
+  @Input() review: Review;
+
+  constructor() {}
+
+  ngOnInit(): void {}
+}