+++ angular/src/app/modules/account/account.module.ts
@@ -1,14 +1,7 @@
 import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-
 import { AccountRoutingModule } from './account-routing.module';
 
-
 @NgModule({
-  declarations: [],
-  imports: [
-    CommonModule,
-    AccountRoutingModule
-  ]
+  imports: [AccountRoutingModule],
 })
-export class AccountModule { }
+export class AccountModule {}