+++ angular/src/app/modules/account/account/account.component.spec.ts
@@ -1,13 +1,51 @@
+import { HttpClientTestingModule } from '@angular/common/http/testing';
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { of } from 'rxjs';
 import { AccountComponent } from './account.component';
+import { AccountBookingComponent } from '../account-booking/account-booking.component';
+import { AccountReviewComponent } from '../account-review/account-review.component';
+import { AddressComponent } from '../address/address.component';
+import { PaymentComponent } from '../payment/payment.component';
+import { ProfileComponent } from '../profile/profile.component';
+import { Account } from '../../../data/account.model';
+import { AccountService } from '../../../services/account/account.service';
 
 describe('AccountComponent', () => {
+  const accountServiceStub = {
+    get() {
+      const account: Account = {
+        id: '',
+        address: {
+          id: '',
+          city: '',
+          country: '',
+          postalCode: '',
+          stateProvince: '',
+          street: '',
+        },
+        payments: null,
+        profiles: null,
+      };
+
+      return of(account);
+    },
+  };
+
   let component: AccountComponent;
   let fixture: ComponentFixture<AccountComponent>;
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      declarations: [AccountComponent],
+      declarations: [
+        AccountComponent,
+        AccountBookingComponent,
+        AccountReviewComponent,
+        AddressComponent,
+        PaymentComponent,
+        ProfileComponent,
+      ],
+      imports: [HttpClientTestingModule],
+      providers: [{ provide: AccountService, useValue: accountServiceStub }],
     }).compileComponents();
   }));