+++ angular/src/app/modules/account/address/address.component.spec.ts
@@ -1,0 +1,33 @@
+import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { AddressComponent } from './address.component';
+
+describe('AddressComponent', () => {
+  const address = {
+    id: '',
+    city: '',
+    country: '',
+    postalCode: '',
+    stateProvince: '',
+    street: '',
+  };
+
+  let component: AddressComponent;
+  let fixture: ComponentFixture<AddressComponent>;
+
+  beforeEach(async(() => {
+    TestBed.configureTestingModule({
+      declarations: [AddressComponent],
+    }).compileComponents();
+  }));
+
+  beforeEach(() => {
+    fixture = TestBed.createComponent(AddressComponent);
+    component = fixture.componentInstance;
+    component.address = address;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});