+++ angular/src/app/modules/account/address/address.component.ts
@@ -1,0 +1,14 @@
+import { Component, Input, OnInit } from '@angular/core';
+import { Address } from '../../../data/address.model';
+
+@Component({
+  selector: 'uic-address',
+  templateUrl: './address.component.html',
+})
+export class AddressComponent implements OnInit {
+  @Input() address: Address;
+
+  constructor() {}
+
+  ngOnInit(): void {}
+}