+++ angular/src/app/modules/account/payment/payment.component.spec.ts
@@ -1,0 +1,30 @@
+import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { PaymentComponent } from './payment.component';
+
+describe('PaymentComponent', () => {
+  const payment = {
+    cardName: '',
+    cardNumber: '',
+    cardExpirationDate: new Date(),
+    id: '',
+  };
+  let component: PaymentComponent;
+  let fixture: ComponentFixture<PaymentComponent>;
+
+  beforeEach(async(() => {
+    TestBed.configureTestingModule({
+      declarations: [PaymentComponent],
+    }).compileComponents();
+  }));
+
+  beforeEach(() => {
+    fixture = TestBed.createComponent(PaymentComponent);
+    component = fixture.componentInstance;
+    component.payment = payment;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});