+++ angular/src/app/modules/account/profile/profile.component.ts
@@ -1,0 +1,14 @@
+import { Component, Input, OnInit } from '@angular/core';
+import { Profile } from '../../../data/profile.model';
+
+@Component({
+  selector: 'uic-profile',
+  templateUrl: './profile.component.html',
+})
+export class ProfileComponent implements OnInit {
+  @Input() profiles: Profile[];
+
+  constructor() {}
+
+  ngOnInit(): void {}
+}